refactor(http): reorganize the code before new features/updates
This commit reorganizes the HttpClient codebase. In particular, it combines
a number of classes/interfaces which previously existed in separate files
into a single `http.ts` file. This is necessary in preparation for the
switch to tree-shakable providers, which will introduce new dependency
arrows that would otherwise result in cycles.
